Removed path-history from Larva.
Larva will not keep track of past state (with the exception of the
history it needs for the kernel calculations).

Views are now responsible for keeping track of any state information
that they depend on. Larva will provide this to them via notifiying
the Model that it has changed, which will in turn update all the views.