Larva now has verbosity, removed duplicated code
Whether a Larva tells the world about its changes or not is a
characteristic of the Larva, so it makes sense for a verbosity
flag to be a member of the Larva class. This will also free us
from having to pass a flag for printing to every function that
wants to print. Now we have a single-point-of-maintenance for
printing.
Verbosity is toggled from the Controller.
Removed 'silent' command.